Make show/hide links nofollow for google SEO
As you can see in this link:

https://www.google.com/search?q=viewtopic_modify_post_list_sql&ei=kJ80XKztF-Sh_Qar-rS4DQ&start=60&sa=N&filter=0&ved=0ahUKEwis9YO-n97fAhXkUN8KHSs9Ddc4MhDw0wMIlQE&biw=1696&bih=763

it seems that google picks up on the show link and crawls that page, adding an ugly page to the sites search results.

If site owners leave it on while the crawler happens to be on the site, these will show up... there's no need for that.

I also edited the acp files and hide links for good measure.